+++ orix/vector/__init__.py
@@ -19,23 +19,27 @@
 """Three-dimensional quantities.
 
 Vectors can represent positions in three-dimensional space and are also
-commonly associated with motion, possessing both a magnitude and a direction.
-In orix they are often encountered as derived objects such as the rotation
-axis of a quaternion or the normal to the bounding planes of a spherical
-region.
+commonly associated with motion, possessing both a magnitude and a
+direction. In orix they are often encountered as derived objects such as
+the rotation axis of a quaternion or the normal to the bounding planes
+of a spherical region.
 """
 
+# Must be imported first
 from orix.vector.vector3d import check_vector, Vector3d
+
+from orix.vector.miller import Miller
 from orix.vector.neo_euler import AxAngle, Homochoric, NeoEuler, Rodrigues
 from orix.vector.spherical_region import SphericalRegion
 
 # Lists what will be imported when calling "from orix.vector import *"
 __all__ = [
     "check_vector",
-    "Vector3d",
     "AxAngle",
     "Homochoric",
+    "Miller",
     "NeoEuler",
     "Rodrigues",
     "SphericalRegion",
+    "Vector3d",
 ]