+++ orix/vector/vector3d.py
@@ -16,8 +16,6 @@
 # You should have received a copy of the GNU General Public License
 # along with orix.  If not, see <http://www.gnu.org/licenses/>.
 
-import warnings
-
 import matplotlib.pyplot as plt
 import numpy as np