+++ CHANGELOG.rst
@@ -12,6 +12,8 @@ Unreleased
 
 Added
 -----
+- Miller class, inherinting functionality from the Vector3d class, to handle operations
+  with direct lattice vectors (uvw/UVTW) and reciprocal lattice vectors (hkl/hkil).
 - Warning when trying to create rotations from large Euler angles
 - Passing symmetry when initializing an Orientation.
 - Vector3d.scatter() and Vector3d.draw_circle() methods to show unit vectors and